WIP reroll nearly-dead cells.
In this version, deadness is Dohare et al 2024's "low utility":

    https://www.nature.com/articles/s41586-024-07711-7

That is, the utility (here we called it "influence") of a cell is
the norm of its outgoing weight covector times an EWMA of the norms
of its activation values. This implementation uses L2 in an attempt
to avoid rerolling cells that encode rare but important details.