Now nearly-constant counts as nearly dead.
Now a cell's influence is its variance times how much it contributes
(in mean square) to the gradient of the objective function, so this
also covers nearly-ignored cells.

(This will probably have to be rewritten for batch normalization.)

When we reroll a nearly-constant cell, adjust the bias terms of its
downstream consumers to compensate.

It's still not clear what the principled way to schedule these rerolls
is; but in order to be less disruptive, switch to only rerolling the
lowest-influence cell, waiting for the minimum age if necessary.